move queue to shared
