+++ pynumaflow/reducer/_dtypes.py
@@ -8,7 +8,7 @@ from typing import TypeVar, Callable, Union
 from collections.abc import AsyncIterable
 from warnings import warn
 
-from pynumaflow.reducer.servicer.asynciter import NonBlockingIterator
+from pynumaflow.shared.asynciter import NonBlockingIterator
 from pynumaflow._constants import DROP
 
 M = TypeVar("M", bound="Message")
